Completed initial set up up to point of adding test files and bootstrap.
