+++ .env
@@ -1,0 +1,2 @@
+USERNAME=username 
+PASSWORD=password
