+++ app/assets/stylesheets/admin.scss
@@ -1,0 +1,1 @@
+@import "bootstrap";
