+++ app/controllers/admin/application_controller.rb
@@ -1,0 +1,3 @@
+class Admin::ApplicationController < ActionController::Base
+  layout "admin"
+end
