+++ app/views/admin/articles/_form.html.erb
@@ -1,4 +1,4 @@
-<%= form_with model: @article, local: true do |form| %>
+<%= form_for [:admin, @article] do |form| %>
  
   <% if @article.errors.any? %>
     <div id="error_explanation">
