+++ app/views/admin/articles/edit.html.erb
@@ -2,4 +2,4 @@
  
 <%= render 'form' %>
  
-<%= link_to 'Back', articles_path %>
+<%= link_to 'Back', admin_articles_path %>
