+++ app/views/admin/articles/index.html.erb
@@ -1,0 +1,23 @@
+<h1>Listing articles</h1>
+ 
+<%= link_to 'New article', new_admin_article_path %>
+
+<table>
+  <tr>
+    <th>Title</th>
+    <th>Text</th>
+    <th></th>
+  </tr>
+ 
+  <% @articles.each do |article| %>
+    <tr>
+      <td><%= article.title %></td>
+      <td><%= article.text %></td>
+      <td><%= link_to 'View', admin_article_path(article) %></td>
+      <td><%= link_to 'Edit', edit_admin_article_path(article) %></td>
+      <td><%= link_to 'Delete', admin_article_path(article),
+              method: :delete,
+              data: { confirm: 'Are you sure?' } %></td>
+      </tr>
+  <% end %>
+</table>
