+++ app/views/admin/articles/show.html.erb
@@ -1,0 +1,21 @@
+<p>
+  <strong>Title:</strong>
+  <%= @article.title %>
+</p>
+ 
+<p>
+  <strong>Text:</strong>
+  <%= @article.text %>
+</p>
+
+<%= link_to "Edit", edit_admin_article_path(@article) %>
+
+<br>
+
+<%= link_to 'Delete', admin_article_path(@article),
+    method: :delete,
+    data: { confirm: 'Are you sure?' } %>
+
+<br>
+
+<%= link_to 'Back to article list', admin_articles_path %>
