+++ app/views/articles/index.html.erb
@@ -1,6 +1,4 @@
 <h1>Listing articles</h1>
- 
-<%= link_to 'New article', new_article_path %>
 
 <table>
   <tr>
@@ -14,12 +12,6 @@
       <td><%= article.title %></td>
       <td><%= article.text %></td>
       <td><%= link_to 'View', article_path(article) %></td>
-      <td><%= link_to 'Edit', edit_article_path(article) %></td>
-      <td><%= link_to 'Delete', article_path(article),
-              method: :delete,
-              data: { confirm: 'Are you sure?' } %></td>
-      </tr>
+    </tr>
   <% end %>
 </table>
-
-<%= link_to 'Back to home page', welcome_index_path %>
