--- app/views/welcome/index.html.erb
@@ -1,4 +1,0 @@
-<h1>Welcome to my Personal Portfolio Site</h1>
-<p>This is a personal portfolio website built with Ruby on Rails.</p>
-<p>The aim being to showcase my coding skills and knowledge.</p>
-<%= link_to 'My Articles', controller: 'articles' %>
