+++ test/controllers/articles_controller_test.rb
@@ -1,7 +1,8 @@
 require 'test_helper'
 
 class ArticlesControllerTest < ActionDispatch::IntegrationTest
-  # test "the truth" do
-  #   assert true
-  # end
+  test "should see articles" do
+    get articles_show_url
+    assert_response :success
+  end
 end
