--- test/controllers/welcome_controller_test.rb
@@ -1,9 +1,0 @@
-require 'test_helper'
-
-class WelcomeControllerTest < ActionDispatch::IntegrationTest
-  test "should get index" do
-    get welcome_index_url
-    assert_response :success
-  end
-
-end
